--- GraspDataProcessing/__init__.py
@@ -1,41 +1,0 @@
-#!/usr/bin/env python
-# -*- encoding: utf-8 -*-
-'''
-@Id :__init__.py
-@date :2023/03/20 14:59:52
-@author :YenochQin (秦毅)
-'''
-import sys
-import warnings
-
-# If a version with git hash was stored, use that instead
-from . import version
-from .version import __version__
-
-import numpy as np
-import pandas as pd
-from pathlib import Path
-import re
-from tqdm import tqdm
-
-
-from .data_IO import *
-from .ASF_data_collection import *
-from .transition_data_collection import *
-from .radial_wavefunction_plot import *
-from .fig_settings import *
-
-
-
-# 根据需要导出部分内容
-__all__ = ['np',
-           'pd',
-           'Path',
-           're',
-           'GraspFileLoad',
-           'EnergyFile2csv',
-           'set_size',
-           'mcdhf_energy_data_collection',
-           'ci_energy_data_collection',
-           'tqdm'
-           ]
