--- GraspDataProcessing/version.py
@@ -1,2 +1,0 @@
-# version.py
-__version__ = '1.1.0'
