+++ pyproject.toml
@@ -1,28 +1,40 @@
-# pyproject.toml is a configuration file for Python projects.
-# It is used to specify dependencies, build system, and other metadata.
-
 [build-system]
-requires = ["setuptools", "wheel"]
-build-backend = "setuptools.build_meta"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
 
-[tool.poetry]
-name = "GraspDataProcessing"
-version = "1.1.0"
-description = "A Python package for processing grasp data."
-authors = [{ name="Yenoch(Yi) Qin", email="Yenochqin@hotmail.com" },]
-license = "GPL"
+[project]
+name = "grasp-data-processing"
+version = "1.1.3"
+description = "Grasp Data processing package"
+
+requires-python = ">= 3.8"
+
+dependencies = [
+  "matplotlib",
+  "numpy",
+  "pandas",
+  "tqdm",
+]
+
 readme = "README.md"
+license = { file = "LICENSE" }
+authors = [
+  { name="Yi(Yenoch) Qin", email="yenochqin@hotmail.com" }
+]
 
+classifiers = [
+  # How mature is this project? Common values are
+  #   3 - Alpha
+  #   4 - Beta
+  #   5 - Production/Stable
+  "Development Status :: 4 - Beta",
 
-[tool.poetry.dependencies]
-python = "^3.8"
-numpy = "^1.20.1"
-pandas = "^1.2.3"
-matplotlib = "^3.3.4"
-re = "*"
-tqdm = "^4.59.0"
-pathlib = "^1.0.1"
-[tool.poetry.dev-dependencies]
-pytest = "^6.2.2"
+  # Specify the Python versions you support here.
+  "Programming Language :: Python",
+]
 
-[tool.poetry.scripts]
+[tool.hatch.build.targets.wheel]
+packages = ["src/graspdataprocessing"]
+
+[project.urls]
+Homepage = "https://github.com/YenochQin/graspdataprocessing"
