+++ requirements.txt
@@ -1,0 +1,4 @@
+matplotlib==3.8.4
+numpy==2.0.0
+pandas==2.2.2
+tqdm==4.66.4
