+++ src/graspdataprocessing/version.py
@@ -1,0 +1,2 @@
+# version.py
+__version__ = '1.1.3'
